refactor: added useDappRedirect.ts
