+++ src/components/dapp-staking/dapp/Dapp.vue
@@ -1,21 +1,21 @@
 <template>
   <div v-if="dapp">
-    <DappAvatar :dapp="dapp" />
-    <DappStatistics :dapp="dapp" />
-    <DappImages :dapp="dapp" />
-    <Builders :dapp="dapp" />
+    <dapp-avatar :dapp="dapp" />
+    <dapp-statistics :dapp="dapp" />
+    <dapp-images :dapp="dapp" />
+    <builders :dapp="dapp" />
     <div class="row--project-overview">
-      <ProjectOverview :dapp="dapp" />
-      <ProjectDetails :dapp="dapp" />
+      <project-overview :dapp="dapp" />
+      <project-details :dapp="dapp" />
     </div>
   </div>
 </template>
 <script lang="ts">
-import { useNetworkInfo, useStakingList } from 'src/hooks';
+import { useNetworkInfo, useStakingList, useDappRedirect } from 'src/hooks';
 import { Path } from 'src/router';
 import { useStore } from 'src/store';
 import { computed, defineComponent, watchEffect } from 'vue';
-import { useRoute, useRouter } from 'vue-router';
+import { useRoute } from 'vue-router';
 import DappAvatar from 'src/components/dapp-staking/dapp/DappAvatar.vue';
 import DappStatistics from 'src/components/dapp-staking/dapp/DappStatistics.vue';
 import DappImages from 'src/components/dapp-staking/dapp/DappImages.vue';
@@ -36,7 +36,7 @@ export default defineComponent({
   setup() {
     const { currentNetworkName } = useNetworkInfo();
     const route = useRoute();
-    const router = useRouter();
+    useDappRedirect();
     const { t } = useI18n();
     const store = useStore();
     const { dapps, stakingList } = useStakingList();
@@ -72,20 +72,6 @@ export default defineComponent({
       }
       return null;
     });
-
-    const handleRedirect = (): void => {
-      if (dappAddress.value && dapps.value.length > 0) {
-        const dapp = dapps.value.find(
-          (it: any) => it.contract.address.toLowerCase() === dappAddress.value.toLowerCase()
-        );
-        !dapp &&
-          router.push({
-            path: Path.DappStaking,
-          });
-      }
-    };
-
-    watchEffect(handleRedirect);
     watchEffect(dispatchGetDapps);
 
     return {
