+++ src/hooks/index.ts
@@ -29,6 +29,7 @@ export * from './dapps-staking/useNominationTransfer';
 export * from './dapps-staking/useStakerInfo';
 export * from './dapps-staking/useStakingList';
 export * from './dapps-staking/useSignPayload';
+export * from './dapps-staking/useDappRedirect';
 export * from './wallet/useWalletIcon';
 export * from './wallet/useEvmWallet';
 export * from './xcm/useXcmBridgeV3';
